percentages can now take floating point values

+++ src/limelight/styles/abstrstyling/PercentageAttribute.java
@@ -2,5 +2,5 @@ package limelight.styles.abstrstyling;
 
 public interface PercentageAttribute extends StyleAttribute
 {
-  int getPercentage();
+  double getPercentage();
 }
