+++ src/limelight/ui/model/PaintJob.java
@@ -98,10 +98,10 @@ public class PaintJob
   public void applyAlphaCompositeFor(Panel panel, Graphics2D graphics)
   {
     Style style = panel.getStyle();
-    int alphaPercentage = style.getCompiledTransparency().getPercentage();
+    double alphaPercentage = style.getCompiledTransparency().getPercentage();
     if(alphaPercentage > 0)
     {
-      float alpha = 1.0f - (alphaPercentage / 100.0f);
+      float alpha = (float)(1.0f - (alphaPercentage / 100.0f));
       Composite alphaComposite = AlphaComposite.getInstance(AlphaComposite.SRC_OVER, alpha);
       graphics.setComposite(alphaComposite);
     }
