+++ src/limelight/ui/painting/GradientPainter.java
@@ -24,7 +24,7 @@ public class GradientPainter extends Painter
     Color color1 = style.getCompiledBackgroundColor().getColor();
     Color color2 = style.getCompiledSecondaryBackgroundColor().getColor();
     int angle = style.getCompiledGradientAngle().getDegrees();
-    int penetration = style.getCompiledGradientPenetration().getPercentage();
+    double penetration = style.getCompiledGradientPenetration().getPercentage();
     boolean cyclic = style.getCompiledCyclicGradient().isOn();
     int x1, y1, x2, y2;
     if(angle == 90)
